+++ BDDTDD/src/main/java/Stepsdefinitions/TMKOCEP.java
@@ -18,16 +18,23 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 
 public class TMKOCEP {
+	public static WebDriver driver;
 	
-	WebDriver driver;
-		
+	static {
+		String browserPath = "/Applications/Brave Browser.app/Contents/MacOS/Brave Browser";
+		ChromeOptions options = new ChromeOptions();
+		options.setBinary(browserPath);
+		driver = new ChromeDriver(options);
+		driver.manage().window().maximize();
+	}
+	
 		@Given("Launch Brave Browser")
-		public void setUpClass() throws InterruptedException, IOException {
-			String browserPath = "/Applications/Brave Browser.app/Contents/MacOS/Brave Browser";
-			ChromeOptions options = new ChromeOptions();
-			options.setBinary(browserPath);
-			driver = new ChromeDriver(options);
-			driver.manage().window().maximize();
+		public void launch_brave_browser() throws InterruptedException, IOException {
+//			String browserPath = "/Applications/Brave Browser.app/Contents/MacOS/Brave Browser";
+//			ChromeOptions options = new ChromeOptions();
+//			options.setBinary(browserPath);
+//			driver = new ChromeDriver(options);
+//			driver.manage().window().maximize();
 			System.out.println("Browser launched Successfully");
 //			screenshot(driver, System.currentTimeMillis());
 			System.out.println("Looking for clearCacheCookes");
@@ -101,12 +108,7 @@ public class TMKOCEP {
 			System.out.println("TMKOC episode searched successfully");
 		}
 
-		@When("Search naukari and open sussessfully")
-		public void search_naukari_and_open_sussessfully() {
-			driver.get("https://www.naukri.com/mnjuser/homepage");
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-			System.out.println("YouTube title verified successfully");
-		}
+		
 		
 		@Then("I enter {string} and {string}")
 		public void i_enter_and(String string, String string2) {
@@ -132,50 +134,13 @@ public class TMKOCEP {
 			driver.quit();
 		}
 		
-		@When("Search Instahyre and open sussessfully")
-		public void search_instahyre_and_open_sussessfully() {
-			driver.get("https://www.instahyre.com/");
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-			driver.findElement(By.xpath("//a[text()='Login']")).click();
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-			System.out.println("InstaHyre title verified successfully");
-		}
 		
-		@Then("I enter {string} and {string} of instahyre")
-		public void i_enter_and_of_instahyre(String string, String string2) {
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-			driver.findElement(By.xpath("//input[@id='email']")).sendKeys(string);
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-			driver.findElement(By.xpath("//input[@id='password']")).sendKeys(string2);
-			driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5000));
-		}
 		
-		@Then("I upload Resume on Naukari")
-		public void i_upload_resume_on_naukari() throws InterruptedException {
-//			String projectPath = System.getProperty("user.dir"); 
-//			System.out.println("ProjectFilePath : "+projectPath);
-//			WebElement updateResumeButton = driver.findElement(By.xpath("//input[@type='button' and @value='Update resume']"));
-//			updateResumeButton.click();
-			WebElement fileInput = driver.findElement(By.xpath("//input[@type='file']"));
-			fileInput.sendKeys("/Users/tarunkishore/git/repository/BDDTDD/src/test/resources/Utilities/Tarunkishore_SQA.pdf");
-		}
 		
-		@Then("I update resume for instahyre")
-		public void i_update_resume_for_instahyre() throws InterruptedException {
-//			String projectPath = System.getProperty("user.dir"); 
-//			System.out.println("ProjectFilePath : "+projectPath);
-//			Thread.sleep(10000);
-//			WebElement element = driver.findElement(By.xpath("//h6[contains(text(),'Education')]"));
-//			JavascriptExecutor js = (JavascriptExecutor) driver;
-//			js.executeScript("arguments[0].scrollIntoView(true);", element);
-//			Thread.sleep(10000);
-			WebElement updateResumeButton = driver.findElement(By.xpath("(//label[@for='resume-input'])[2]"));
-			updateResumeButton.click();
-			Thread.sleep(10000);
-			WebElement fileInput = driver.findElement(By.xpath("//input[@type='file']"));
-			fileInput.sendKeys("/Users/tarunkishore/git/repository/BDDTDD/src/test/resources/Utilities/Tarunkishore_SQA.pdf");
-			Thread.sleep(10000);	
-		}
+		
+		
+		
+		
 		
 		@When("Search foundit and open sussessfully")
 		public void search_foundit_and_open_sussessfully() {
